Refactor: Decompose HandBar and PromptBubbleEditor below 50-line limit
