+++ eslint.config.mjs
@@ -153,13 +153,11 @@ export default ts.config(
       "src/components/organisms/EasyModeView.tsx",
       "src/components/organisms/EvolutionSuccessModal.tsx",
       "src/components/organisms/ExpertModeView.tsx",
-      "src/components/organisms/HandBar.tsx",
       "src/components/organisms/HistoryTab.tsx",
       "src/components/organisms/InteractiveTutorial.tsx",
       "src/components/organisms/LocalBackupSection.tsx",
       "src/components/organisms/MergeStackModal.tsx",
       "src/components/organisms/OnboardingGuide.tsx",
-      "src/components/organisms/PromptBubbleEditor.tsx",
       "src/components/organisms/SettingsTab.tsx",
       "src/components/organisms/ShareCardModal.tsx",
       "src/components/organisms/SimpleMintingView.tsx",
